Draggable grouping has landed in jCanvas, as well as many new methods and fixes.
NEW
- Draggable groups have been implemented in this latest release. Simply
set a layer's 'dragGroupWithLayer' property to true. As a result,
dragging that layer will move all layers in the same layer group. For
examples, see the documentation.
- Added a new setLayers() method for easily setting properties of all
layers (rather than a specific layer)
- The setLayer() method can now be used to assign events at any point
after a layer's creation. Before, adding events could only be done when
initially creating a layer
- Added a new getLayerIndex() method for accurately retrieving a
layer's index
- Added the animated layer as a third argument to the callback function
for the animateLayer() method
- The getLayers() method now accepts an optional function argument
which will be used as a callback to filter the layers array
- The animateLayer() method now allows you to use functions as values
for any of the properties to be animated

CHANGED
- The drawGraph() method has been deprecated

FIXED
- Fixed a bug which prevented layer events from firing if an image
layer was present
- Fixed a bug in which clearing part of the canvas using clearCanvas()
did not respect transformation properties (rotate, scale, etc.)