Fix bug with drag event canvas propagation
Originally, the extra check I added ensured that dragging on a lower
canvas was not interrupted on a canvas above. However, that only worked
if the lower canvas was adjacent to the canvas above. It would not work
if both canvases were separated by a third canvas. This commit adds a
global data store to enable jCanvas to keep track of the
currently-dragged layer across all canvases at a given point in time.