Limit redraws when user is not dragging a layer
When a layer is marked as draggable, jCanvas will internally bind a
mousemove event to the canvas. Ideally, that mousemove event should
trigger a canvas redraw *only if* a layer is being dragged or if the
user has explicitly bound a mouseover/mouseout/mousemove event to some
layer. However, for the longest time, jCanvas would still needlessly
redraw the canvas in this case even if a layer was not being dragged.
This commit corrects that longstanding inefficiency by *only* redrawing
if a user event is triggered or if a layer is being dragged, but in no
other case.